+++ CHANGELOG.md
@@ -62,6 +62,8 @@ All notable changes to the Yumma CSS will be documented in this file.
 ### Added
 
 - Add `$yma-box-shadow-*` variables
+- Add `yummacss.config.js` file
+- Add Yumma CSS CLI
 
 ### Changes